requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
openai==1.25.1
python-dotenv==1.0.1
pydantic==2.7.1
streamlit==1.33.0
requests==2.28.2
gunicorn==20.1.0
